serve: refactor functionality into papis.web
